requirements.txt
aiogram
requests
bs4
python-binance